+++ test/blast/test_alignments.py
@@ -13,10 +13,9 @@ try:
 except ImportError:
     from mock import patch
 
-from six import StringIO
 from Bio import SeqIO
 
-from ..mocking import mockOpen, File
+from ..mocking import mockOpen
 from .sample_data import PARAMS, RECORD0, RECORD1, RECORD2, RECORD3, RECORD4
 
 from dark.reads import Read, Reads, DNARead
@@ -26,6 +25,7 @@ from dark.blast.alignments import (
     BlastReadsAlignments, ZERO_EVALUE_UPPER_RANDOM_INCREMENT)
 from dark.titles import TitlesAlignments
 from dark import ncbidb
+from dark.utils import StringIO
 
 
 class TestBlastReadsAlignments(TestCase):
@@ -214,7 +214,8 @@ class TestBlastReadsAlignments(TestCase):
         If a JSON file contains a parameters section and one record, it must
         be read correctly.
         """
-        result = File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+        result = StringIO(dumps(PARAMS) + '\n' +
+                          dumps(RECORD0) + '\n')
 
         with patch.object(builtins, 'open') as mockMethod:
             mockMethod.return_value = result
@@ -237,9 +238,11 @@ class TestBlastReadsAlignments(TestCase):
             def sideEffect(self, _ignoredFilename, **kwargs):
                 if self.first:
                     self.first = False
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD0) + '\n')
                 else:
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD1) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD1) + '\n')
 
         sideEffect = SideEffect()
         with patch.object(builtins, 'open') as mockMethod:
@@ -268,9 +271,11 @@ class TestBlastReadsAlignments(TestCase):
             def sideEffect(self, _ignoredFilename, **kwargs):
                 if self.first:
                     self.first = False
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD2) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD2) + '\n')
                 else:
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD3) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD3) + '\n')
 
         sideEffect = SideEffect()
         with patch.object(builtins, 'open') as mockMethod:
@@ -305,14 +310,17 @@ class TestBlastReadsAlignments(TestCase):
                 if self.count == 0:
                     self.test.assertEqual('1.json', filename)
                     self.count += 1
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD0) + '\n')
                 elif self.count == 1:
                     self.test.assertEqual('2.json', filename)
                     self.count += 1
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD1) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD1) + '\n')
                 else:
                     self.test.assertEqual('3.json', filename)
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD2) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD2) + '\n')
 
         sideEffect = SideEffect(self)
         with patch.object(builtins, 'open') as mockMethod:
@@ -347,11 +355,13 @@ class TestBlastReadsAlignments(TestCase):
             def sideEffect(self, _ignoredFilename):
                 if self.first:
                     self.first = False
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD0) + '\n')
                 else:
                     params = deepcopy(PARAMS)
                     params['application'] = 'Skype'
-                    return File([dumps(params) + '\n', dumps(RECORD1) + '\n'])
+                    return StringIO(dumps(params) + '\n' +
+                                    dumps(RECORD1) + '\n')
 
         sideEffect = SideEffect()
         with patch.object(builtins, 'open') as mockMethod:
@@ -414,10 +424,11 @@ class TestBlastReadsAlignments(TestCase):
                 if self.count == 0:
                     self.test.assertEqual('file.json', filename)
                     self.count += 1
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD0) + '\n')
                 elif self.count == 1:
                     self.count += 1
-                    return File(['>id1 Description', 'AA\n'])
+                    return StringIO('>id1 Description\nAA\n')
                 else:
                     self.fail('Unexpected third call to open.')
 
@@ -480,11 +491,12 @@ class TestBlastReadsAlignments(TestCase):
                 if self.count == 0:
                     self.test.assertEqual('file.json', filename)
                     self.count += 1
-                    return File([dumps(PARAMS) + '\n', dumps(RECORD0) + '\n'])
+                    return StringIO(dumps(PARAMS) + '\n' +
+                                    dumps(RECORD0) + '\n')
                 elif self.count == 1:
                     self.test.assertEqual('xxx.fasta', filename)
                     self.count += 1
-                    return File(['>seqid\n', 'AA\n'])
+                    return StringIO('>seqid\nAA\n')
                 else:
                     self.fail('Unexpected third call to open.')
 
@@ -551,12 +563,12 @@ class TestBlastReadsAlignments(TestCase):
         evalues are converted to the positive value of their negative exponent.
         """
         def result(a, **kwargs):
-            return File([
-                dumps(PARAMS) + '\n',
-                dumps(deepcopy(RECORD0)) + '\n',
-                dumps(deepcopy(RECORD1)) + '\n',
-                dumps(deepcopy(RECORD2)) + '\n',
-                dumps(deepcopy(RECORD3)) + '\n'])
+            return StringIO(
+                dumps(PARAMS) + '\n' +
+                dumps(deepcopy(RECORD0)) + '\n' +
+                dumps(deepcopy(RECORD1)) + '\n' +
+                dumps(deepcopy(RECORD2)) + '\n' +
+                dumps(deepcopy(RECORD3)) + '\n')
 
         with patch.object(builtins, 'open') as mockMethod:
             mockMethod.side_effect = result
@@ -580,13 +592,13 @@ class TestBlastReadsAlignments(TestCase):
         evalues are set randomly high.
         """
         def result(a, **kwargs):
-            return File([
-                dumps(PARAMS) + '\n',
-                dumps(deepcopy(RECORD0)) + '\n',
-                dumps(deepcopy(RECORD1)) + '\n',
-                dumps(deepcopy(RECORD2)) + '\n',
-                dumps(deepcopy(RECORD3)) + '\n',
-                dumps(deepcopy(RECORD4)) + '\n'])
+            return StringIO(
+                dumps(PARAMS) + '\n' +
+                dumps(deepcopy(RECORD0)) + '\n' +
+                dumps(deepcopy(RECORD1)) + '\n' +
+                dumps(deepcopy(RECORD2)) + '\n' +
+                dumps(deepcopy(RECORD3)) + '\n' +
+                dumps(deepcopy(RECORD4)) + '\n')
 
         with patch.object(builtins, 'open') as mockMethod:
             mockMethod.side_effect = result