+++ test/test_utils.py
@@ -11,7 +11,7 @@ try:
 except ImportError:
     from mock import patch
 
-from .mocking import mockOpen, File
+from .mocking import mockOpen
 
 from dark.utils import (
     numericallySortFilenames, median, asHandle, parseRangeString, StringIO,
@@ -155,7 +155,7 @@ class TestAsHandle(TestCase):
         # This test should be better. It should actually create some bz2
         # compressed data and make sure that it's decompressed
         # properly. But Python mocking makes me so confused...
-        result = File('xxx')
+        result = StringIO('xxx')
 
         with patch.object(bz2, 'BZ2File') as mockMethod:
             mockMethod.return_value = result
@@ -173,7 +173,7 @@ class TestAsHandle(TestCase):
         # This test should be better. It should actually create some gzip
         # compressed data and make sure that it's decompressed
         # properly. But Python mocking makes me so confused...
-        result = File('xxx')
+        result = StringIO('xxx')
 
         with patch.object(gzip, 'GzipFile') as mockMethod:
             mockMethod.return_value = result